requirements.txt
fastapi==0.111.0
Pillow==10.4.0
torch==2.3.1
transformers==4.42.3
numpy==1.26.4
